Moves TestLoadBalancerConfig to top level to shareˆ
