+++ spring-cloud-gateway-server-mvc/src/test/java/org/springframework/cloud/gateway/server/mvc/config/GatewayMvcPropertiesBeanDefinitionRegistrarTests.java
@@ -28,15 +28,10 @@ import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.cloud.client.DefaultServiceInstance;
-import org.springframework.cloud.gateway.server.mvc.ServerMvcIntegrationTests;
+import org.springframework.cloud.gateway.server.mvc.test.TestLoadBalancerConfig;
 import org.springframework.cloud.gateway.server.mvc.test.client.TestRestClient;
 import org.springframework.cloud.loadbalancer.annotation.LoadBalancerClient;
-import org.springframework.cloud.loadbalancer.core.ServiceInstanceListSupplier;
-import org.springframework.cloud.loadbalancer.support.ServiceInstanceListSuppliers;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpMethod;
 import org.springframework.test.context.ActiveProfiles;
@@ -137,7 +132,7 @@ public class GatewayMvcPropertiesBeanDefinitionRegistrarTests {
 
 	@SpringBootConfiguration
 	@EnableAutoConfiguration
-	@LoadBalancerClient(name = "testservice", configuration = ServerMvcIntegrationTests.TestLoadBalancerConfig.class)
+	@LoadBalancerClient(name = "testservice", configuration = TestLoadBalancerConfig.class)
 	static class Config {
 
 	}
@@ -246,19 +241,6 @@ public class GatewayMvcPropertiesBeanDefinitionRegistrarTests {
 		@Override
 		public void unknown(RouterFunction<?> routerFunction) {
 
-		}
-
-	}
-
-	public static class TestLoadBalancerConfig {
-
-		@LocalServerPort
-		protected int port = 0;
-
-		@Bean
-		public ServiceInstanceListSupplier staticServiceInstanceListSupplier() {
-			return ServiceInstanceListSuppliers.from("testservice",
-					new DefaultServiceInstance("testservice" + "-1", "testservice", "localhost", port, false));
 		}
 
 	}
